refactor: asis.Controller 제거
